Embeddings fix (#655)
* track embeddings now

* move get_database call to not load tensorflow

* remove test code

* fail; wrongly imported db from embeddings instead of search

* more fixes

* remove unused import
