+++ birdnet_analyzer/search/core.py
@@ -1,5 +1,6 @@
 from typing import Literal
 
+
 def search(
     output: str,
     database: str,
@@ -38,7 +39,7 @@ def search(
 
     import birdnet_analyzer.audio as audio
     import birdnet_analyzer.config as cfg
-    from birdnet_analyzer.search.utils import get_database, get_search_results
+    from birdnet_analyzer.search.utils import get_search_results
 
     # Create output folder
     if not os.path.exists(output):
@@ -70,3 +71,9 @@ def search(
         sig, rate = audio.open_audio_file(file, offset=offset, duration=duration, sample_rate=None)
         result_path = os.path.join(output, f"{file[4]:.5f}_{filebasename}_{offset}_{offset + duration}.wav")
         audio.save_signal(sig, result_path, rate)
+
+
+def get_database(database_path):
+    from perch_hoplite.db import sqlite_usearch_impl
+
+    return sqlite_usearch_impl.SQLiteUsearchDB.create(database_path).thread_split()
